feat(builder): initial builder work
Creates the initial builder resource and webhooks.  This also removes
the janky cert generators that I was using previously in favor of using
cewrt-manager (which is part of the localdev env).  Also removes the
startup of the kind cluster since it will be managed by localdev.

TODO: need to go back through and make the mount path for the dev
container configurable.