feat(builder): retrieve token
When observing the builder, retrieve the token secret based on the
spec value or use 'coral-github-token'. Needed to make sure that we
were using a pointer to a string for the secret name and also remove
the dev target in favor of install.  Will be wiring up the watch process
tomorrow.