fix(nodeselectors): ensure old selectors are removed
This fixes the bug introduced last night on the initial commit of the
injectors where the managed selectors would not clean themselves up.  In
the case where an image would be updated, the old selector would be
orphaned and not be removed.  This also simplifies and consolidates some
quick and dirty filtering work that was happening for the
include/exclude.  It's more clear what's happening and more idiomatic.