feat(auth): support registry authentication
Adds registry authentication to the agents.  Pull secrets are loaded in
and a keyring and auth configs are created and passed on to the worker
processes.  This ended up being more of a significant refactor of the
original node worker and there's still a bit more cleanup to do in the
polling, but it's a bit cleaner than before.  In conjunction with the
additions, I also reworked the way that I was handling tests.  Instead
of relying on envtest and spinning up a full environment, I wrapped the
controller runtime fake client to support the loading of fixtures and am
now passing the mock client around.  I'll most likely keep envtest
around for integration testing, but all of the unit testing will be
moving to the mock client.