fix(finalizer): ensure finalizers are removed
There were some inconsistent behaviors with how the finalizers were
being managed.  I moved the logic back into the reconciler for now,
fixed some of the issues that were present and put testing around it.
Ther are still issues with the monitor shutting down, but those are not
controller related and I'll address those tomorrow.