fix(monitor): fix monitor shutdown
Fixes the issues with the monitor shutdown, cleans it up a bit, and
reorders some logs and checks in the controller.  I just ran into an
issue where I hit docker's rate limit.  I'll want to make sure that I
address that situation a little better (and handle the situation). Seems
like this would be a good scenario for the local registry.