feat(metrics): initial metrics
First pass at some initial custom metrics for monitoring the state of
the controller and agent.  The monitor has not been fully instrumented
as there are some structural changes that will be following.  In
addition, because the agent does not block for pulls/removals the actual
count for the metrics is wrong.  Updates for some sort of blocking
mechanism are on the way.