fix(monitor): fix monitor status/metrics
This fix actually provides the correct and useful information about the
object.  It also decouples the monitors from the controller so that it
will independently fetch all of the managed images without needing to
add or remove monitors during reconciliation.  This also changes the
status fields to align with images instead of nodes and limits the
number of times the api needs to be queried.  I can probably use the
same logic in the finalizer.