feat(agent): blocking operations for pull/remove
Adding blocking operations.  Query the image management service and get
a list of images instead of querying node status.  This ensures that we
can get the information as quickly as possible and take pressure off of
the api.  The initial change was in the worker, but realized that there
is a lag between the worker and the node picking up the new labels, so
updated the agent loop to pull in the node images from the same source.
To ensure that the agent loop was also blocking and would not add
additional events while we were still processing long running
operations, I moved to an unbuffered channel.  This also gives us the
ability to throttle pulls and removes.