chore(docker): set up docker build
Sets up the dockerfile.  Because we are interacting with the containerd
socket and we need to reduce any sort of attack surface, we just pull in
the essentials and have a very slim container with only certs and the
coral executable.  Also adds a stage overlay to test out the containers
after they are built.