chore(clean): minor cleaning
Cleaned up the pull method for the agent workers and updated some of the
logic there.  There's an issue with the agent placing events on the
channel while pulls are still processing on the workers.  I'm not quite
sure what to do to get around that yet, but the behavior/cause has been
confirmed.  It may take some sort of shared lock.  I also made a note of
a race that was discovered in the monitor worker.  In the current form,
there's a possibility of the finalizer updates taking place after the
monitor has retrieved the object.